move swarm code into quickjs and create skill module backend
