+++ libs/providers/quickjs/src/skills/swarm/SKILL.md
