+++ libs/providers/quickjs/src/skills/swarm/batching.ts
