+++ libs/providers/quickjs/src/skills/swarm/executor.test.ts
