+++ libs/providers/quickjs/src/skills/swarm/executor.ts
