+++ libs/providers/quickjs/src/skills/swarm/filter.ts
