+++ libs/providers/quickjs/src/skills/swarm/index.test.ts
