+++ libs/providers/quickjs/src/skills/swarm/index.ts
