+++ libs/providers/quickjs/src/skills/swarm/interpolate.test.ts
