+++ libs/providers/quickjs/src/skills/swarm/interpolate.ts
