+++ libs/providers/quickjs/src/skills/swarm/table.test.ts
