+++ libs/providers/quickjs/src/skills/swarm/table.ts
