+++ libs/providers/quickjs/src/skills/swarm/types.ts
