+++ libs/providers/quickjs/src/skills/swarm/utils.ts
